Reuser result buffers
Up to now this was already done, but the current code did not re-use existing
windows. So the current buffer was kicked out of the window and was replaced
by the result buffer.

Use switchbuf to jump to a open window in the current tab or open a new one.

Fixes #40